Develop to Stage - Fix no WebAdmin

+++ app/Http/Controllers/RelatorioFotograficoController.php
@@ -75,10 +75,15 @@ class RelatorioFotograficoController extends AppBaseController
     public function deleteRelatorioFotos($id)
     {
         $programacao = $this->programacaoRepository->find($id);
-        $programacao->relatorioFotografico()->delete();
 
-        Flash::success('Relatório excluído com sucesso');
+        if ($programacao->relatorioFotografico) {
+            \File::delete($programacao->relatorioFotografico->pathArquivo);
+            $programacao->relatorioFotografico()->delete();
+            Flash::success('Relatório excluído com sucesso');
+        } else {
+            Flash::error('Não há relatório para ser excluído');
+        }
 
-        return view('programacoes.show')->with('programacao', $programacao);
+        return redirect('programacoes/'.$id);
     }
 }
